Begin Transition to PPU SDRAM DMA
Currently, none of the changes below break the demo or alter it in
any visible way.

Separated VRAM module into CPU-Facing and PPU-Facing variants. The
main difference is that both ports of the CPU-Facing VRAM have a
data-width of 128-bits to maximize compatibility with the
DMA-Engine.

Modified h2f_vram_interface. Its purpose is now to connect the Read
Master port of the DMA-Engine to the CPU-Facing VRAM.

Added a new ppu_dma_interface. Its purpose is to expose the DMA
Control Register interface to our ppu_dma_logic.

Modified vram_sync_writer to utilize the new 128-bit ports of the
CPU-Facing VRAM. This is untested as of yet, but will be tested
shortly. It is unecessary for the demo, as the PPU-Facing RAM is
now the one to be initialized.

Split the vram_if into two separate files, one for CPU-Facing and
one for PPU-Facing VRAMs. Removed the byteen signals from the
interfaces, which are no longer being used.

Modified vram_interconnect. Now CPU-Facing VRAM has a static
structure. Port-A only connects to vram_syncwriter; Port-B only
connects to the h2f_vram_interface.

Fixed a bug with the PPU FSM where if the DMA transfer was late,
the ppu_logic would never receive the row-swap signal, and
therefore would not render frames. The intended behaviour is simply
to avoid the sync state when the DMA transfer is late, and continue
rendering the previous frame as normal.

Add new dma engine Platform Designer IP. This was necessary due to
a bug similar to the following link (but currently undocumented):
https://www.intel.com/content/www/us/en/programmable/support/
support-resources/knowledge-base/tools/2020/why-there-is-a-
synthesis-error-when-there-are-2-instances-of-gen.html.