Split Repo; Add Documentation (Unfinished)
Removed ctests. We can add these back in if necessary for the
user.

Removed the img_src folder. This will be included in the source
repository.

Renamed the_mall_test to techdemo.

Added a docs folder. Currently two important pieces of
documentation are missing. I will be adding these soon.

Added several brief README.md files to locations where I thought
they could be useful, including in the various subfolders of the
example techdemo. The README files included in the techdemo file
need to be finished still. I will be doing this soon

Revised the various .gitignores to ignored files based on the
subfolder. Most notably, a general version of the sdk.mk file is
included now.